Factored out Child tests and a few others
