+++ temporal-sdk/src/test/java/io/temporal/workflow/ChildWorkflowAsyncRetryTest.java
@@ -1,0 +1,124 @@
+/*
+ *  Copyright (C) 2020 Temporal Technologies, Inc. All Rights Reserved.
+ *
+ *  Copyright 2012-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ *  Modifications copyright (C) 2017 Uber Technologies, Inc.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not
+ *  use this file except in compliance with the License. A copy of the License is
+ *  located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file. This file is distributed on
+ *  an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ *  express or implied. See the License for the specific language governing
+ *  permissions and limitations under the License.
+ */
+
+package io.temporal.workflow;
+
+import static org.junit.Assert.*;
+
+import io.temporal.client.WorkflowException;
+import io.temporal.client.WorkflowOptions;
+import io.temporal.common.RetryOptions;
+import io.temporal.failure.ApplicationFailure;
+import io.temporal.failure.ChildWorkflowFailure;
+import io.temporal.testing.TracingWorkerInterceptor;
+import io.temporal.testing.WorkflowReplayer;
+import io.temporal.workflow.shared.SDKTestWorkflowRule;
+import io.temporal.workflow.shared.TestWorkflows;
+import java.time.Duration;
+import org.junit.Assume;
+import org.junit.Rule;
+import org.junit.Test;
+
+public class ChildWorkflowAsyncRetryTest {
+
+  private final WorkflowTest.AngryChildActivityImpl angryChildActivity =
+      new WorkflowTest.AngryChildActivityImpl();
+
+  @Rule
+  public SDKTestWorkflowRule testWorkflowRule =
+      SDKTestWorkflowRule.newBuilder()
+          .setWorkflowTypes(
+              TestChildWorkflowAsyncRetryWorkflow.class, WorkflowTest.AngryChild.class)
+          .setActivityImplementations(angryChildActivity)
+          .setWorkerInterceptors(
+              new TracingWorkerInterceptor(new TracingWorkerInterceptor.FilteredTrace()))
+          .build();
+
+  @Test
+  public void testChildWorkflowAsyncRetry() {
+    WorkflowOptions options =
+        WorkflowOptions.newBuilder()
+            .setWorkflowRunTimeout(Duration.ofSeconds(20))
+            .setWorkflowTaskTimeout(Duration.ofSeconds(2))
+            .setTaskQueue(testWorkflowRule.getTaskQueue())
+            .build();
+    TestWorkflows.TestWorkflow1 client =
+        testWorkflowRule
+            .getWorkflowClient()
+            .newWorkflowStub(TestWorkflows.TestWorkflow1.class, options);
+    try {
+      client.execute(testWorkflowRule.getTaskQueue());
+      fail("unreachable");
+    } catch (WorkflowException e) {
+      assertTrue(String.valueOf(e.getCause()), e.getCause() instanceof ChildWorkflowFailure);
+      assertTrue(e.getCause().getCause() instanceof ApplicationFailure);
+      assertEquals("test", ((ApplicationFailure) e.getCause().getCause()).getType());
+      assertEquals(
+          "message='simulated failure', type='test', nonRetryable=false",
+          e.getCause().getCause().getMessage());
+    }
+    assertEquals(3, angryChildActivity.getInvocationCount());
+  }
+
+  /** Tests that WorkflowReplayer fails if replay does not match workflow run. */
+  @Test(expected = RuntimeException.class)
+  public void testAlteredWorkflowReplayFailure() throws Exception {
+    Assume.assumeFalse("skipping for docker tests", SDKTestWorkflowRule.useExternalService);
+
+    WorkflowReplayer.replayWorkflowExecutionFromResource(
+        "testChildWorkflowRetryHistory.json", AlteredTestChildWorkflowRetryWorkflow.class);
+  }
+
+  public static class TestChildWorkflowAsyncRetryWorkflow implements TestWorkflows.TestWorkflow1 {
+
+    private WorkflowTest.ITestChild child;
+
+    public TestChildWorkflowAsyncRetryWorkflow() {}
+
+    @Override
+    public String execute(String taskQueue) {
+      ChildWorkflowOptions options =
+          ChildWorkflowOptions.newBuilder()
+              .setWorkflowRunTimeout(Duration.ofSeconds(5))
+              .setWorkflowTaskTimeout(Duration.ofSeconds(2))
+              .setTaskQueue(taskQueue)
+              .setRetryOptions(
+                  RetryOptions.newBuilder()
+                      .setMaximumInterval(Duration.ofSeconds(1))
+                      .setInitialInterval(Duration.ofSeconds(1))
+                      .setMaximumAttempts(3)
+                      .build())
+              .build();
+      child = Workflow.newChildWorkflowStub(WorkflowTest.ITestChild.class, options);
+      return Async.function(child::execute, taskQueue, 0).get();
+    }
+  }
+
+  public static class AlteredTestChildWorkflowRetryWorkflow
+      extends TestChildWorkflowAsyncRetryWorkflow {
+
+    public AlteredTestChildWorkflowRetryWorkflow() {}
+
+    @Override
+    public String execute(String taskQueue) {
+      Workflow.sleep(Duration.ofMinutes(1));
+      return super.execute(taskQueue);
+    }
+  }
+}
